Object generator for S4 class spPlot

Object spPlot calls one of the following functions based on the value of initMode:spiralize::spiral_initialize,spiralize::spiral_initialize_by_gcoor,spiralize::spiral_initialize_by_time.
The correct way to call it is as follows:
spPlot(initMode = 'initialize',clear = TRUE,xlim = c(0,1),start = 360,end = 360 *5,scale_by = c("angle","curve_length"),period = NULL,clockwise = FALSE,flip = c("none","vertical","horizontal","both"),reverse = FALSE,polar_lines = scale_by=="angle",polar_lines_by = 30,polar_lines_gp = gpar(col = "#808080"),padding = unit(5,"mm"),newpage = TRUE,vp_param = list())

spPlot(initMode = 'initialize_by_gcoor',clear = TRUE,xlim = NULL,scal_by = "curve_length",...)

spPlot(initMode = 'initialize_by_time',clear = TRUE,xlim = NULL,start = NULL,end = NULL,unit_on_axis = c("days","months","weeks","hours","mins","secs"),period = c("years","months","weeks","days","hours","mins"),normalize_year = FALSE,period_per_loop = 1,polar_lines_by = NULL,verbose = TRUE,...)

Usage

spPlot(initMode = "initialize", clear = TRUE, ...)

Arguments

initMode

It can only be the following values:"initialize", "initialize_by_gcoor", "initialize_by_time".

clear

Whether to call spiralize::spiral_clear before drawing.

...

Arguments passed on to spiralize::spiral_initialize, spiralize::spiral_initialize_by_gcoor, spiralize::spiral_initialize_by_time

xlim

Range on x-locations.

start

Start of the spiral, in degree. start and end should be positive and start should be smaller than end.

end

End of the spiral, in degree.

scale_by

How scales on x-axis are equally interpolated? The values can be one of "angle" and "curve_length". If the value is "angle", equal angle difference corresponds to equal difference of data. In this case, in outer loops, the scales are longer than in the inner loops, although the difference on the data are the same. If the value is "curve_length", equal curve length difference corresponds to the equal difference of the data.

period

Under "angle" mode, the number of loops can also be controlled by argument period which controls the length of data a spiral loop corresponds to. Note in this case, argument end is ignored and the value for end is internally recalculated.

clockwise

Whether the curve is in a closewise direction. If it is set to TRUE, argument flip and reverse are ignored.

flip

How to flip the spiral? By default, the spiral starts from the origin of the coordinate and grows reverseclockwisely. The argument controls the growing direction of the spiral.

reverse

By default, the most inside of the spiral corresponds to the lower boundary of x-location. Setting the value to FALSE can reverse the direction.

polar_lines

Whether draw the polar guiding lines.

polar_lines_by

Increment of the polar lines. Measured in degree. The value can also be a vector that defines where to add polar lines.

polar_lines_gp

Graphics parameters for the polar lines.

padding

Padding of the plotting region. The value can be a grid::unit() of length of one to two.

newpage

Whether to apply grid::grid.newpage() before making the plot?

vp_param

A list of parameters sent to grid::viewport().

unit_on_axis

Units on the axis.

normalize_year

Whether to enforce one spiral loop to represent a complete year?

period_per_loop

How many periods to put in a spiral loop?

verbose

Whether to print messages?

Value

Object spPlot

Examples

NULL
#> NULL