Define a generic track

Object spTrack will call the function spiralize::spiral_track while drawing.

Usage

spTrack(
  ylim = c(0, 1),
  height = 0.8,
  background = TRUE,
  background_gp = gpar(fill = "#EEEEEE"),
  reverse_y = FALSE,
  gradient = FALSE
)

Arguments

ylim

Data range of the y-locations.

height

Height of the track. The value can be the fraction of the distance of the two neighbour spiral loops. The value can also be a grid::unit() object.

background

Whether to draw the background of the track, i.e. border and filled color of background.

background_gp

Graphical parameters of the background.

reverse_y

Whether reverse the direction of y-axis (i.e. pointing to the center of the spiral)?

gradient

Whether draw the background in gradient? The value can be a positive integer of the number of gradients from background_gp$fill to white.

Value

Object spTrack

Examples

NULL
#> NULL